common/vision/lasr_vision_lang_sam/requirements.txt
aiofiles==24.1.0          # via gradio
annotated-types==0.7.0    # via pydantic
antlr4-python3-runtime==4.9.3  # via hydra-core, omegaconf
anyio==4.9.0              # via gradio, httpx, starlette, watchfiles
certifi==2025.6.15        # via httpcore, httpx, requests
charset-normalizer==3.4.2  # via requests
click==8.2.1              # via typer, uvicorn
contourpy==1.3.2          # via matplotlib, supervision
cycler==0.12.1            # via matplotlib
defusedxml==0.7.1         # via supervision
exceptiongroup==1.3.0     # via anyio
fastapi==0.115.13         # via gradio, litserve
ffmpy==0.6.0              # via gradio
filelock==3.18.0          # via huggingface-hub, torch, transformers, triton
fonttools==4.58.4         # via matplotlib
fsspec==2025.5.1          # via gradio-client, huggingface-hub, torch
gradio==5.29.0            # via -r requirements.in, lang-sam
gradio-client==1.10.0     # via gradio
groovy==0.1.2             # via gradio
h11==0.16.0               # via httpcore, uvicorn
hf-xet==1.1.4             # via huggingface-hub
httpcore==1.0.9           # via httpx
httptools==0.6.4          # via uvicorn
httpx==0.28.1             # via gradio, gradio-client, safehttpx
huggingface-hub==0.33.0   # via gradio, gradio-client, tokenizers, transformers
hydra-core==1.3.2         # via sam-2
idna==3.10                # via anyio, httpx, requests
iopath==0.1.10            # via sam-2
jinja2==3.1.6             # via gradio, torch
kiwisolver==1.4.8         # via matplotlib
lang-sam @ git+https://github.com/LASR-at-Home/lang-segment-anything.git@main  # via -r requirements.in
litserve==0.2.8           # via -r requirements.in, lang-sam
markdown-it-py==3.0.0     # via rich
markupsafe==3.0.2         # via gradio, jinja2
matplotlib==3.10.3        # via supervision
mdurl==0.1.2              # via markdown-it-py
mpmath==1.3.0             # via sympy
networkx==3.4.2           # via torch
numpy==2.2.6              # via contourpy, gradio, matplotlib, opencv-python, opencv-python-headless, pandas, sam-2, scipy, supervision, torchvision, transformers
nvidia-cublas-cu12==12.1.3.1  # via nvidia-cudnn-cu12, nvidia-cusolver-cu12, torch
nvidia-cuda-cupti-cu12==12.1.105  # via torch
nvidia-cuda-nvrtc-cu12==12.1.105  # via torch
nvidia-cuda-runtime-cu12==12.1.105  # via torch
nvidia-cudnn-cu12==9.1.0.70  # via torch
nvidia-cufft-cu12==11.0.2.54  # via torch
nvidia-curand-cu12==10.3.2.106  # via torch
nvidia-cusolver-cu12==11.4.5.107  # via torch
nvidia-cusparse-cu12==12.1.0.106  # via nvidia-cusolver-cu12, torch
nvidia-nccl-cu12==2.20.5  # via torch
nvidia-nvjitlink-cu12==12.9.86  # via nvidia-cusolver-cu12, nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105  # via torch
omegaconf==2.3.0          # via hydra-core
opencv-python==4.11.0.86  # via supervision
opencv-python-headless==4.10.0.84  # via -r requirements.in, lang-sam
orjson==3.10.18           # via gradio
packaging==25.0           # via gradio, gradio-client, huggingface-hub, hydra-core, matplotlib, transformers
pandas==2.3.0             # via gradio
pillow==11.2.1            # via gradio, matplotlib, sam-2, supervision, torchvision
portalocker==3.2.0        # via iopath
pydantic==2.11.7          # via -r requirements.in, fastapi, gradio, lang-sam
pydantic-core==2.33.2     # via pydantic
pydub==0.25.1             # via gradio
pygments==2.19.1          # via rich
pyparsing==3.2.3          # via matplotlib
python-dateutil==2.9.0.post0  # via matplotlib, pandas
python-dotenv==1.1.0      # via uvicorn
python-multipart==0.0.20  # via gradio
pytz==2025.2              # via pandas
pyyaml==6.0.2             # via gradio, huggingface-hub, omegaconf, supervision, transformers, uvicorn
pyzmq==27.0.0             # via litserve
regex==2024.11.6          # via transformers
requests==2.32.4          # via huggingface-hub, supervision, transformers
rich==14.0.0              # via typer
ruff==0.12.0              # via gradio
safehttpx==0.1.6          # via gradio
safetensors==0.5.3        # via transformers
sam-2 @ git+https://github.com/facebookresearch/segment-anything-2@c2ec8e14a185632b0a5d8b161928ceb50197eddc  # via -r requirements.in, lang-sam
scipy==1.15.3             # via supervision
semantic-version==2.10.0  # via gradio
shellingham==1.5.4        # via typer
six==1.17.0               # via python-dateutil
sniffio==1.3.1            # via anyio
starlette==0.46.2         # via fastapi, gradio
supervision==0.25.1       # via -r requirements.in, lang-sam
sympy==1.14.0             # via torch
tokenizers==0.19.1        # via transformers
tomlkit==0.13.3           # via gradio
torch==2.4.1              # via -r requirements.in, lang-sam, sam-2, torchvision
torchvision==0.19.1       # via -r requirements.in, lang-sam, sam-2
tqdm==4.67.1              # via huggingface-hub, iopath, sam-2, supervision, transformers
transformers==4.44.2      # via -r requirements.in, lang-sam
triton==3.0.0             # via torch
typer==0.16.0             # via gradio
typing-extensions==4.14.0  # via anyio, exceptiongroup, fastapi, gradio, gradio-client, huggingface-hub, iopath, pydantic, pydantic-core, rich, torch, typer, typing-inspection, uvicorn
typing-inspection==0.4.1  # via pydantic
tzdata==2025.2            # via pandas
urllib3==2.5.0            # via requests
uvicorn[standard]==0.34.3  # via gradio, litserve
uvloop==0.20.0            # via -r requirements.in, lang-sam, uvicorn
watchfiles==1.1.0         # via uvicorn
websockets==15.0.1        # via gradio-client, uvicorn
